# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ticket Dashboard",
    "description": "A lightweight ticketing REST API with intentionally unimplemented features to demonstrate GitHub MCP, Jira MCP, and TestRail MCP workflows.",
    "contact": {
      "name": "Ticket Dashboard Team"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/integrations/jira/sync/{ticket_id}": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Sync ticket to Jira (planned)",
        "description": "Placeholder endpoint for future Jira synchronization. Returns 501 by design.",
        "operationId": "sync_ticket_to_jira_integrations_jira_sync__ticket_id__post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sync Ticket To Jira Integrations Jira Sync  Ticket Id  Post"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/testrail/push/{ticket_id}": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Push ticket to TestRail (planned)",
        "description": "Placeholder endpoint for future TestRail synchronization. Returns 501 by design.",
        "operationId": "push_ticket_to_testrail_integrations_testrail_push__ticket_id__post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Push Ticket To Testrail Integrations Testrail Push  Ticket Id  Post"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/github/create-issue/{ticket_id}": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Create GitHub issue from ticket (planned)",
        "description": "Placeholder endpoint for future GitHub issue creation. Returns 501 by design.",
        "operationId": "create_github_issue_for_ticket_integrations_github_create_issue__ticket_id__post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Github Issue For Ticket Integrations Github Create Issue  Ticket Id  Post"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}